ADCP Utilities

This is a collection of functions that are helpful for analysing ADCP data. Some of them are specifically for pulse-coherent ADCPs, but some are more generally useful.

General

DataMethods.get_sig_fnames — Function

get_sig_fnames(filepath, file_prefix)

Get the ordered list of filepaths for Signature NetCDF file with given file_prefix in filepath.

Signal Analysis

These functions relate to analysing the pulse and acoustic signal itself for pulse coherent ADCPs.

DataMethods.calculate_SNR — Function

calculate_SNR(As, An; mode=:log)

Calculate the signal-to-noise ratio of signal amplitude As relative to noise amplitude An.

Specify that As and An are in :log-space or :lin-space with mode.

DataMethods.HR_V_variance — Function

HR_V_variance(c, ds)

Calculate the velocity variance of sample of correlation c and parameters in NCDataset ds.

HR_V_variance(c, Δr, F₀, ΔF, τ; cs = 1500)

Calculate the velocity variance of sample of correlation c.

...

Arguments

Δr: cell size.

F₀: center frequency.

ΔF: frequency bandwidth.

τ: time lag between pings in a ping pair.

cs: speed of sound.

...

DataMethods.HR_c_from_V_variance — Function

HR_c_from_V_variance(var, ds)

Calculate the velocity correlation of sample of variance var and parameters in ds.

HR_c_from_V_variance(var, Δr, F₀, ΔF, τ; cs = 1500)

Calculate the velocity correlation of sample of variance var.

...

Arguments

Δr: cell size.

F₀: center frequency.

ΔF: frequency bandwidth.

τ: time lag between pings in a ping pair.

cs: speed of sound.

...

DataMethods.PulseLag — Function

PulseLag(λ, c)

Calculate the pulse lag τ from lag distance λ and speed of sound c.

DataMethods.IntrinsicVelocityRange — Function

IntrinsicVelocityRange(c, F₀, τ)

Calculate the intrinsic velocity range for a pulse-coherent ADCP with speed of sound c, carrier frequency F₀, and pulse lag τ.

DataMethods.ExtendedVelocityRange — Function

ExtendedVelocityRange(c, F₀, τ; nmax=6)

Calculate the extended velocity range for a pulse-coherent ADCP with speed of sound c, carrier frequency F₀, pulse lag τ, and the number of side correlations considered nmax.

DataMethods.CorrelationFromTurbulence — Function

CorrelationFromTurbulence(ϵ, d, Vᵣ; C₁=0.707)

Calculate the maximum correlation for a pulse-coherent ADCP expected in a region with TKE dissipation rate ϵ.

...

Arguments

ϵ: TKE dissipation rate (m² s⁻³)

d: characteristic measurement size (largest of either cell length or width) (m)

Vᵣ: intrinsic velocity range (m s⁻¹).

...

DataMethods.TurbulenceFromCorrelation — Function

TurbulenceFromCorrelation(cor, d, Vᵣ; C₁=0.707)

Calculate the TKE dissipation rate that would result in correlation c for a pulse-coherent ADCP.

...

Arguments

cor: correlation

d: characteristic measurement size (largest of either cell length or width) (m)

Vᵣ: intrinsic velocity range (m s⁻¹).

...

Transformations

DataMethods.rm_sidelobe_contam — Function

rm_sidelobe_contam(data, datarange, boundary)

Remove the side-lobe-contaminated measurements from data.

The maximum radial datarange the ADCP can measure is equal to the distance from the ADCP to the boundary. The contaminated measurements are therefore at along-beam dataranges greater than or equal to boundary.

...

Arguments

data: the data to decontaminate.

datarange: the along beam distance from the ADCP. ADCP attitude does not affect it, and it is the same for the slant beams.

boundary: the distance to the visible boundary (i.e. the boundary sound is reflecting off). This is also independent of ADCP attitude.

...

Coordinate Tranformations

DataMethods.align_z0 — Function

align_z0(data, zs, z0, Δz; buffer=15)

Shift the columns in data and zs such that the point closest to z0 in zs is aligned in a row.

This is useful for shifting bottom-referenced data to surface-referenced and vice versa, for example.

...

Arguments

data: the data to shift.

zs: the coordinates of data (must be size(data)).

z0: the value in zs to use for aligning.

Δz: the range step to use to construct a reference coordinate vector.

buffer: an arbitrary buffer to add to the size shifted array to ensure all the data fits (must be greater than the largest shift). The arrays are reduced to the correct size after.

...

DataMethods.r_2_slant_r — Function

r_2_slant_r(range, θ)

Convert ADCP nominal range or cell size to the equivalent for slant beams at angle θ (in deg) to beam 5.

DataMethods.z_2_brange — Function

z_2_brange(z, θ, ϕ, roll, pitch)

Convert z-coordinate to along beam range using the ADCP roll and pitch attitude.

θ is the slant beam angle in the ADCP x-direction from z and ϕ is the same in the y-direction.

DataMethods.brange_2_z — Function

brange_2_z(brange, θ, ϕ, roll, pitch)

Convert along-beam range to z coordinate using the ADCP roll and pitch attitude.

θ is the slant beam angle in the ADCP x-direction from z and ϕ is the same in the y-direction.

Velocity Transformations

Developed for the Signature 1000, but useful more generally.

DataMethods.transformation_matrix — Function

transformation_matrix(θ)

Make the transformation matrix for a 4-beam ADCP with beam angle θ in degrees.

DataMethods.heading_R_matrix — Function

heading_R_matrix(heading)

Make the heading rotation matrix for a 4-beam ADCP with magnetometer reading heading.

DataMethods.pitchroll_R_matrix — Function

pitchroll_R_matrix(roll, pitch)

Make the pitch/roll rotation matrix for a 4-beam ADCP with attitude readings roll and pitch.

DataMethods.rotation_matrix — Function

rotation_matrix(AHRSraw)

Make the full rotation matrix for a 4-beam ADCP using the data from an on-board AHRS sensor.

rotation_matrix(heading, pitch, roll)

Make the full rotation matrix for a 4-beam ADCP using the attitude data (in degrees) heading, pitch, and roll.

DataMethods.beam2xyz — Function

beam2xyz(V1, V2, V3, V4; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to xyz-referenced velocities. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

DataMethods.beam2ENU — Function

beam2ENU(V1, V2, V3, V4, AHRSraw; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to earth-referenced velocities using the data from an AHRS sensor. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

beam2ENU(V1, V2, V3, V4, heading, pitch, roll; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to earth-referenced velocities using the attitude data. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

Developed for the RDI Workhorse 600

DataMethods.RDI_pitch2pitch — Function

RDI_pitch2pitch(pitch, roll)

Convert the RDI instrument measured pitch to the gimbal pitch required for the rotation matrix.

Units are in degrees.

DataMethods.RDI_rotation_matrix — Function

RDI_rotation_matrix(pitch, roll, heading)

Make the rotation matrix for an RDI instrument defined with pitch, roll, and heading.

Units are in degrees.

DataMethods.RDI_ENU2XYZ — Function

RDI_ENU2XYZ(uE, vN, U, pitch, roll, heading)

Convert earth-referenced to instrument coordinate velocities for RDI data.

Angle units are in degrees.

DataMethods.RDI_XYZ2ENU — Function

RDI_XYZ2ENU(u, v, w, pitch, roll, heading)

Convert instrument coordinates to earth-referenced for RDI data.

Angle units are in degrees.

Wave Analysis

This function is for wave analysis from the motion of a floating ADCP. The method is described in Aucoin et al. 2026 (under review).

DataMethods.wave_U_ofz — Function

wave_U_ofz(IA, IF, IP, z, h, θ, β, α; method=:velocity, incl_trend=false, IMFs2incl=1:size(IA)[2], g = 9.80665)

Calculate the instantaneous surface gravity wave-induced along-beam velocity.

The matrix algebra only works one way so be careful what shape the inputs are. All the trigonometry and geometry assumes the ADCP is negligably far from the surface. The projections would be off if the distance is significant.

...

Arguments

IA: instantaneous amplitude from the Hilbert-Huang transform.

IF: instantaneous frequency from the Hilbert-Huang transform.

IP: instantaneous phase from the Hilbert-Huang transform.

r: along-beam range coordinate (from the surface).

h: bottom depth.

θ: angle of the beam relative to vertical toward the beam.

θ₀: fixed beam angle.

β: angle of the beam relative to vertical perpendicular the beam.

α: angle between the wave vector and the beam axis projected onto the surface plane.

method: what property to base the amplitude on. Options are :velocity and :acceleration.

incl_trend=true: optionally include the mean trend of the IMF decomposition.

IMFs2incl: the mode numbers to include in the calculation.

g: optionally specify the acceleration due to gravity.

u0: initial wavenumber guess for the numerical solver.

...