CTD Utilities

These are functions that are generally useful for analysing CTD data.

DataMethods.find_ctd_profiles — Function

find_ctd_profiles(pressure; sm = 65, Pthresh = 1, stdthresh = 4, nomotion_tol = 0.05)

Find down and upcasts from the pressure timeseries of a CTD (developed for the ecoCTD).

...

Arguments

pressure: time series of pressure.

sm: smoothing factor for the pressure derivative.

Pthresh: threshold of maximum pressure for the "surface".

stdthresh: factor of the standard deviation of the derivative of pressure that is used to determine if the CTD is profiling.

nomotion_tol: tolerance factor for how close to 0 the speed has to be to be considered stopped.

...

DataMethods.correct_CT_lag — Function

correct_CT_lag(T, fs; τ = 0.35)

Shift the temperature T forward in time by τ seconds to account for the inherent lag between conductivity and temperature measurements for RBR CTDs.

The approach and default value of 0.35 seconds are presented in Dever et al. (2022).

DataMethods.remove_loops — Function

remove_loops(data; tol = 0, max_z = 0, replace=false)

Remove all values in the fields of data that have a higher z value than the previous lowest.

This is equivalent to removing points that do not monotonically decrease in z. Set tol to add a tolerance to the strict z < z_min. Optionally set a starting max_z to something other than 0.

If replace=true then the the data points are replaced with NaN instead of being removed.

DataMethods.FindLoops — Function

FindLoops(p; tol = 0, max_z = 0)

Find all indices in p that are not monotonically increasing.

Optionally specify a tolerance tol and a starting max_z.