Library

All functions exported by this package are listed below.

Library Contents

Library

Library Contents

Index

Functions

Calculus Utilities

Signal Processing

ADCP Utilities

CTD Utilities

GPS Utilities

Index

DataMethods.CorrelationFromTurbulence

DataMethods.ExtendedVelocityRange

DataMethods.FT_params

DataMethods.FindLoops

DataMethods.HR_V_variance

DataMethods.HR_c_from_V_variance

DataMethods.IntervalAverage

DataMethods.IntrinsicVelocityRange

DataMethods.PulseLag

DataMethods.RDI_ENU2XYZ

DataMethods.RDI_XYZ2ENU

DataMethods.RDI_pitch2pitch

DataMethods.RDI_rotation_matrix

DataMethods.TurbulenceFromCorrelation

DataMethods.Welch

DataMethods.add_ms

DataMethods.align_z0

DataMethods.beam2ENU

DataMethods.beam2xyz

DataMethods.brange_2_z

DataMethods.calculate_SNR

DataMethods.clean_and_mean

DataMethods.clean_gps_data

DataMethods.correct_CT_lag

DataMethods.despike

DataMethods.find_ctd_profiles

DataMethods.findspikes

DataMethods.get_sig_fnames

DataMethods.heading_R_matrix

DataMethods.index_by_time

DataMethods.int_def_trap

DataMethods.int_trap

DataMethods.interp_2d

DataMethods.mean_gps_positions

DataMethods.pitchroll_R_matrix

DataMethods.r_2_slant_r

DataMethods.remove_loops

DataMethods.rfactor2x

DataMethods.rfactor2z

DataMethods.rm_sidelobe_contam

DataMethods.rotation_matrix

DataMethods.transformation_matrix

DataMethods.unit_spherical_vector_projection_u

DataMethods.unit_spherical_vector_projection_w

DataMethods.wave_U_ofz

DataMethods.wave_from_acceleration

DataMethods.wave_from_velocity

DataMethods.z_2_brange

DataMethods.∂_back

DataMethods.∂_cen

Functions

DataMethods.IntervalAverage — Function

IntervalAverage(data::AbstractMatrix{T} where T<:Any, Δx, fs, D)

Average data in chunks of physical dimensional size Δx over its second dimension. fs is the sample frequency of the second dimension, and has inverse units of Δx. D is the fraction of the interval Δx that actually exists in data (the duty cycle fraction).

The purpose of this function is to take a dataset that has discontinous jumps at intervals of size D*Δx, add implied "no data" of length (1-D)*Δx, then average over interval Δx. This function replaces lowertresmean.

IntervalAverage(data::AbstractMatrix{T} where T<:Any, Δx, fs, D, δ₁, δ₂)

Average data in chunks of physical dimensional size Δx over its second dimension. fs is the sample frequency of the second dimension, and has inverse units of Δx. D is the fraction of the interval Δx that actually exists in data (the duty cycle fraction). δ₁ and δ₂ are fractions of D for which there is data for the first and last chunks respectively. This is to account for situations where the start or end don't have a complete duty cycle.

The purpose of this function is to take a dataset that has discontinous jumps at intervals of size D*Δx, add implied "no data" of length (1-D)*Δx, then average over interval Δx. This function replaces lowertresmean.

DataMethods.interp_2d — Function

interp_2d(arrayjl)

Interpolate array in 2 dimensions using a 2D cubic spline fit.

Uses scipy to make my life easier.

Calculus Utilities

DataMethods.∂_cen — Function

∂_cen(fx, x)

A simple centered difference estimation of the derivative of fx with respect to axis x.

DataMethods.∂_back — Function

∂_back(fx, x)

A simple backward finite difference estimation of the derivative of fx with respect to axis x.

DataMethods.int_def_trap — Function

int_def_trap(fx, Δx)

A simple trapezoidal method for estimating a definite integral of fx with separation Δx.

DataMethods.int_trap — Function

int_trap(fx, x)

A simple trapezoidal method for estimating an indefinite integral of fx with axis x.

Signal Processing

DataMethods.Welch — Function

Welch(segments, N; win=nothing)

Compute the power spectrum using Welch's method and periodogram of the input time series segments.

...

Arguments

segments: must be a tuple of vectors of the segments to FFT, even if there is only one segment.

N: the number of elements in each segment.

win: optional window to apply to each segment.

...

DataMethods.FT_params — Function

FT_params(fs, s_seg; fil_frac = 0, win_s = :rect, ovlap = 0.5)

Create the parameters for computing and plotting a Welch periodogram.

...

Arguments

fs: Sample frequency of the data.

s_seg: Welch window length (in inverse units of fs).

fil_frac: Smoothing window size in % of data.

win_s: Type of Welch taper window function.

ovlap: Fractional amount of overlap between Welch segments.

...

DataMethods.findspikes — Function

findspikes(data; sm=35, std_thresh=2)

A simple method to find spikes in data.

The moving average of data is computed as a reference ref_data. Values in data that are more than std_thresh standard deviations of data different from ref_data are replaced with the corresponding ref_data point.

DataMethods.despike — Function

despike(data; sm=35, std_thresh=2)

A simple method to despike data.

The moving average of data is computed as a reference ref_data. Values in data that are more than std_thresh standard deviations of data different from ref_data are replaced with the corresponding ref_data point.

ADCP Utilities

DataMethods.get_sig_fnames — Function

get_sig_fnames(filepath, file_prefix)

Get the ordered list of filepaths for Signature NetCDF file with given file_prefix in filepath.

Signal Analysis

DataMethods.calculate_SNR — Function

calculate_SNR(As, An; mode=:log)

Calculate the signal-to-noise ratio of signal amplitude As relative to noise amplitude An.

Specify that As and An are in :log-space or :lin-space with mode.

DataMethods.HR_V_variance — Function

HR_V_variance(c, ds)

Calculate the velocity variance of sample of correlation c and parameters in NCDataset ds.

HR_V_variance(c, Δr, F₀, ΔF, τ; cs = 1500)

Calculate the velocity variance of sample of correlation c.

...

Arguments

Δr: cell size.

F₀: center frequency.

ΔF: frequency bandwidth.

τ: time lag between pings in a ping pair.

cs: speed of sound.

...

DataMethods.HR_c_from_V_variance — Function

HR_c_from_V_variance(var, ds)

Calculate the velocity correlation of sample of variance var and parameters in ds.

HR_c_from_V_variance(var, Δr, F₀, ΔF, τ; cs = 1500)

Calculate the velocity correlation of sample of variance var.

...

Arguments

Δr: cell size.

F₀: center frequency.

ΔF: frequency bandwidth.

τ: time lag between pings in a ping pair.

cs: speed of sound.

...

DataMethods.PulseLag — Function

PulseLag(λ, c)

Calculate the pulse lag τ from lag distance λ and speed of sound c.

DataMethods.IntrinsicVelocityRange — Function

IntrinsicVelocityRange(c, F₀, τ)

Calculate the intrinsic velocity range for a pulse-coherent ADCP with speed of sound c, carrier frequency F₀, and pulse lag τ.

DataMethods.ExtendedVelocityRange — Function

ExtendedVelocityRange(c, F₀, τ; nmax=6)

Calculate the extended velocity range for a pulse-coherent ADCP with speed of sound c, carrier frequency F₀, pulse lag τ, and the number of side correlations considered nmax.

DataMethods.CorrelationFromTurbulence — Function

CorrelationFromTurbulence(ϵ, d, Vᵣ; C₁=0.707)

Calculate the maximum correlation for a pulse-coherent ADCP expected in a region with TKE dissipation rate ϵ.

...

Arguments

ϵ: TKE dissipation rate (m² s⁻³)

d: characteristic measurement size (largest of either cell length or width) (m)

Vᵣ: intrinsic velocity range (m s⁻¹).

...

DataMethods.TurbulenceFromCorrelation — Function

TurbulenceFromCorrelation(cor, d, Vᵣ; C₁=0.707)

Calculate the TKE dissipation rate that would result in correlation c for a pulse-coherent ADCP.

...

Arguments

cor: correlation

d: characteristic measurement size (largest of either cell length or width) (m)

Vᵣ: intrinsic velocity range (m s⁻¹).

...

Wave Analysis

DataMethods.wave_U_ofz — Function

wave_U_ofz(IA, IF, IP, z, h, θ, β, α; method=:velocity, incl_trend=false, IMFs2incl=1:size(IA)[2], g = 9.80665)

Calculate the instantaneous surface gravity wave-induced along-beam velocity.

The matrix algebra only works one way so be careful what shape the inputs are. All the trigonometry and geometry assumes the ADCP is negligably far from the surface. The projections would be off if the distance is significant.

...

Arguments

IA: instantaneous amplitude from the Hilbert-Huang transform.

IF: instantaneous frequency from the Hilbert-Huang transform.

IP: instantaneous phase from the Hilbert-Huang transform.

r: along-beam range coordinate (from the surface).

h: bottom depth.

θ: angle of the beam relative to vertical toward the beam.

θ₀: fixed beam angle.

β: angle of the beam relative to vertical perpendicular the beam.

α: angle between the wave vector and the beam axis projected onto the surface plane.

method: what property to base the amplitude on. Options are :velocity and :acceleration.

incl_trend=true: optionally include the mean trend of the IMF decomposition.

IMFs2incl: the mode numbers to include in the calculation.

g: optionally specify the acceleration due to gravity.

u0: initial wavenumber guess for the numerical solver.

...

Unexported

DataMethods.wave_from_velocity — Function

wave_from_velocity(A, ω, ϕ, r, h, k, θ, β, α)

Calculate the projected along-beam velocity from a linear surface gravity wave with amplitude A in terms of velocity.

...

Arguments

A: amplitude in terms of velocity.

ω: angular frequency.

ϕ: wave phase.

r: along-beam range coordinate (from the surface).

h: bottom depth.

θ: angle of the beam relative to vertical toward the beam.

θ₀: fixed beam angle.

β: angle of the beam relative to vertical perpendicular the beam.

α: angle between the wave vector and the beam axis projected onto the surface plane.

...

DataMethods.wave_from_acceleration — Function

wave_from_acceleration(A, ω, ϕ, r, h, k, θ, β, α)

Calculate the projected along-beam velocity from a linear surface gravity wave with amplitude A in terms of acceleration.

...

Arguments

A: amplitude in terms of acceleration.

ω: angular frequency.

ϕ: wave phase.

r: along-beam range coordinate (from the surface).

h: bottom depth.

θ: angle of the beam relative to vertical toward the beam.

θ₀: fixed beam angle.

β: angle of the beam relative to vertical perpendicular the beam.

α: angle between the wave vector and the beam axis projected onto the surface plane.

...

DataMethods.unit_spherical_vector_projection_u — Function

unit_spherical_vector_projection_u(θ, α, β)

Utility function to define the scale factor for projection of the horizontal wave component.

DataMethods.unit_spherical_vector_projection_w — Function

unit_spherical_vector_projection_w(θ, β)

Utility function to define the scale factor for projection of the vertical wave component.

DataMethods.rfactor2x — Function

rfactor2x(θ, θ₀, β, α)

Utility function to define the scale factor for the projection of the ADCP range onto the wave x-axis.

DataMethods.rfactor2z — Function

rfactor2z(θ, θ₀, β)

Utility function to define the scale factor for the projection of the ADCP range onto the wave z-axis.

Transformations

DataMethods.rm_sidelobe_contam — Function

rm_sidelobe_contam(data, datarange, boundary)

Remove the side-lobe-contaminated measurements from data.

The maximum radial datarange the ADCP can measure is equal to the distance from the ADCP to the boundary. The contaminated measurements are therefore at along-beam dataranges greater than or equal to boundary.

...

Arguments

data: the data to decontaminate.

datarange: the along beam distance from the ADCP. ADCP attitude does not affect it, and it is the same for the slant beams.

boundary: the distance to the visible boundary (i.e. the boundary sound is reflecting off). This is also independent of ADCP attitude.

...

Coordinate Tranformations

DataMethods.align_z0 — Function

align_z0(data, zs, z0, Δz; buffer=15)

Shift the columns in data and zs such that the point closest to z0 in zs is aligned in a row.

This is useful for shifting bottom-referenced data to surface-referenced and vice versa, for example.

...

Arguments

data: the data to shift.

zs: the coordinates of data (must be size(data)).

z0: the value in zs to use for aligning.

Δz: the range step to use to construct a reference coordinate vector.

buffer: an arbitrary buffer to add to the size shifted array to ensure all the data fits (must be greater than the largest shift). The arrays are reduced to the correct size after.

...

DataMethods.r_2_slant_r — Function

r_2_slant_r(range, θ)

Convert ADCP nominal range or cell size to the equivalent for slant beams at angle θ (in deg) to beam 5.

DataMethods.z_2_brange — Function

z_2_brange(z, θ, ϕ, roll, pitch)

Convert z-coordinate to along beam range using the ADCP roll and pitch attitude.

θ is the slant beam angle in the ADCP x-direction from z and ϕ is the same in the y-direction.

DataMethods.brange_2_z — Function

brange_2_z(brange, θ, ϕ, roll, pitch)

Convert along-beam range to z coordinate using the ADCP roll and pitch attitude.

θ is the slant beam angle in the ADCP x-direction from z and ϕ is the same in the y-direction.

Velocity Transformations

DataMethods.transformation_matrix — Function

transformation_matrix(θ)

Make the transformation matrix for a 4-beam ADCP with beam angle θ in degrees.

DataMethods.heading_R_matrix — Function

heading_R_matrix(heading)

Make the heading rotation matrix for a 4-beam ADCP with magnetometer reading heading.

DataMethods.pitchroll_R_matrix — Function

pitchroll_R_matrix(roll, pitch)

Make the pitch/roll rotation matrix for a 4-beam ADCP with attitude readings roll and pitch.

DataMethods.rotation_matrix — Function

rotation_matrix(AHRSraw)

Make the full rotation matrix for a 4-beam ADCP using the data from an on-board AHRS sensor.

rotation_matrix(heading, pitch, roll)

Make the full rotation matrix for a 4-beam ADCP using the attitude data (in degrees) heading, pitch, and roll.

DataMethods.beam2xyz — Function

beam2xyz(V1, V2, V3, V4; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to xyz-referenced velocities. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

DataMethods.beam2ENU — Function

beam2ENU(V1, V2, V3, V4, AHRSraw; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to earth-referenced velocities using the data from an AHRS sensor. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

beam2ENU(V1, V2, V3, V4, heading, pitch, roll; T=nothing, θ=nothing)

Convert 4 beam-coordinate velocities V1, V2, V3, V4 to earth-referenced velocities using the attitude data. Uses either a supplied transformation matrix T, or makes one from the beam angle θ.

DataMethods.RDI_pitch2pitch — Function

RDI_pitch2pitch(pitch, roll)

Convert the RDI instrument measured pitch to the gimbal pitch required for the rotation matrix.

Units are in degrees.

DataMethods.RDI_rotation_matrix — Function

RDI_rotation_matrix(pitch, roll, heading)

Make the rotation matrix for an RDI instrument defined with pitch, roll, and heading.

Units are in degrees.

DataMethods.RDI_ENU2XYZ — Function

RDI_ENU2XYZ(uE, vN, U, pitch, roll, heading)

Convert earth-referenced to instrument coordinate velocities for RDI data.

Angle units are in degrees.

DataMethods.RDI_XYZ2ENU — Function

RDI_XYZ2ENU(u, v, w, pitch, roll, heading)

Convert instrument coordinates to earth-referenced for RDI data.

Angle units are in degrees.

CTD Utilities

DataMethods.find_ctd_profiles — Function

find_ctd_profiles(pressure; sm = 65, Pthresh = 1, stdthresh = 4, nomotion_tol = 0.05)

Find down and upcasts from the pressure timeseries of a CTD (developed for the ecoCTD).

...

Arguments

pressure: time series of pressure.

sm: smoothing factor for the pressure derivative.

Pthresh: threshold of maximum pressure for the "surface".

stdthresh: factor of the standard deviation of the derivative of pressure that is used to determine if the CTD is profiling.

nomotion_tol: tolerance factor for how close to 0 the speed has to be to be considered stopped.

...

DataMethods.correct_CT_lag — Function

correct_CT_lag(T, fs; τ = 0.35)

Shift the temperature T forward in time by τ seconds to account for the inherent lag between conductivity and temperature measurements for RBR CTDs.

The approach and default value of 0.35 seconds are presented in Dever et al. (2022).

DataMethods.remove_loops — Function

remove_loops(data; tol = 0, max_z = 0, replace=false)

Remove all values in the fields of data that have a higher z value than the previous lowest.

This is equivalent to removing points that do not monotonically decrease in z. Set tol to add a tolerance to the strict z < z_min. Optionally set a starting max_z to something other than 0.

If replace=true then the the data points are replaced with NaN instead of being removed.

DataMethods.FindLoops — Function

FindLoops(p; tol = 0, max_z = 0)

Find all indices in p that are not monotonically increasing.

Optionally specify a tolerance tol and a starting max_z.

GPS Utilities

DataMethods.add_ms — Function

add_ms(gps_time; fs_gps = 10)

Add milliseconds to repeated seconds values in array.

DataMethods.clean_gps_data — Function

clean_gps_data(gps_pth::String; time_format = "YYYYmmddHHMMSS", tz = tz"UTC", fs_gps = 10)

Add milliseconds to GPS time with only seconds and add missing for data that are missing and save as a file.

DataMethods.mean_gps_positions — Function

mean_gps_positions(gps_pths)

Average the latitude and longitude of multiple GPSs and save as a file.

DataMethods.clean_and_mean — Function

clean_and_mean(gps_pths::Vector{String}; time_format = "YYYYmmddHHMMSS", tz = tz"UTC", fs_gps = 10, datatypes=[Int, Bool, String, String, String, String, Int, Float64, Int])

Clean the GPS data up and average the latitude and longitude of multiple GPSs and save each as a file.

DataMethods.index_by_time — Function

index_by_time(times, start_times, end_times)

Find the indices to separate times based on start_times and end_times.