Miscellaneous

These are some miscellaneous data analysis functions.

General

DataMethods.IntervalAverage — Function

IntervalAverage(data::AbstractMatrix{T} where T<:Any, Δx, fs, D)

Average data in chunks of physical dimensional size Δx over its second dimension. fs is the sample frequency of the second dimension, and has inverse units of Δx. D is the fraction of the interval Δx that actually exists in data (the duty cycle fraction).

The purpose of this function is to take a dataset that has discontinous jumps at intervals of size D*Δx, add implied "no data" of length (1-D)*Δx, then average over interval Δx. This function replaces lowertresmean.

IntervalAverage(data::AbstractMatrix{T} where T<:Any, Δx, fs, D, δ₁, δ₂)

Average data in chunks of physical dimensional size Δx over its second dimension. fs is the sample frequency of the second dimension, and has inverse units of Δx. D is the fraction of the interval Δx that actually exists in data (the duty cycle fraction). δ₁ and δ₂ are fractions of D for which there is data for the first and last chunks respectively. This is to account for situations where the start or end don't have a complete duty cycle.

The purpose of this function is to take a dataset that has discontinous jumps at intervals of size D*Δx, add implied "no data" of length (1-D)*Δx, then average over interval Δx. This function replaces lowertresmean.

DataMethods.interp_2d — Function

interp_2d(arrayjl)

Interpolate array in 2 dimensions using a 2D cubic spline fit.

Uses scipy to make my life easier.

Calculus Utilities

DataMethods.∂_cen — Function

∂_cen(fx, x)

A simple centered difference estimation of the derivative of fx with respect to axis x.

DataMethods.∂_back — Function

∂_back(fx, x)

A simple backward finite difference estimation of the derivative of fx with respect to axis x.

DataMethods.int_def_trap — Function

int_def_trap(fx, Δx)

A simple trapezoidal method for estimating a definite integral of fx with separation Δx.

DataMethods.int_trap — Function

int_trap(fx, x)

A simple trapezoidal method for estimating an indefinite integral of fx with axis x.

Signal Processing

DataMethods.Welch — Function

Welch(segments, N; win=nothing)

Compute the power spectrum using Welch's method and periodogram of the input time series segments.

...

Arguments

segments: must be a tuple of vectors of the segments to FFT, even if there is only one segment.

N: the number of elements in each segment.

win: optional window to apply to each segment.

...

DataMethods.FT_params — Function

FT_params(fs, s_seg; fil_frac = 0, win_s = :rect, ovlap = 0.5)

Create the parameters for computing and plotting a Welch periodogram.

...

Arguments

fs: Sample frequency of the data.

s_seg: Welch window length (in inverse units of fs).

fil_frac: Smoothing window size in % of data.

win_s: Type of Welch taper window function.

ovlap: Fractional amount of overlap between Welch segments.

...

DataMethods.findspikes — Function

findspikes(data; sm=35, std_thresh=2)

A simple method to find spikes in data.

The moving average of data is computed as a reference ref_data. Values in data that are more than std_thresh standard deviations of data different from ref_data are replaced with the corresponding ref_data point.

DataMethods.despike — Function

despike(data; sm=35, std_thresh=2)

A simple method to despike data.

The moving average of data is computed as a reference ref_data. Values in data that are more than std_thresh standard deviations of data different from ref_data are replaced with the corresponding ref_data point.